Code cleanup; split some SplitPdfHook methods to separate files
