+++ src/hooks/custom/utils/general.ts
@@ -1,4 +1,3 @@
 export function stringToBoolean(string: string): boolean {
   return string.toLocaleLowerCase() === "true";
 }
-
